+++ Changelog.md
@@ -2,6 +2,8 @@ Flan 1.7.10
 ================
 - Check for keep inventory when deciding if drops should be locked or not  
   Stops unlock cmd message when keep inventory is on
+- Fix addClaim cmd bypassing world blacklist
+- Make admin mode ignore world blacklist
 
 Flan 1.7.9
 ================